refactored section and fixed title
