+++ apps/app/src/components/TopCompaniesSection/TopCompaniesSection.module.scss
