+++ apps/app/src/components/TopCompaniesSection/index.ts
