+++ apps/app/src/pages/CompaniesPage/CompaniesTab/CompaniesTab.tsx
@@ -1,5 +1,5 @@
 import LocationSection from '@/components/LocationSection';
-import TopCompaniesSection from '@/pages/Home/sections/TopCompaniesSection';
+import TopCompaniesSection from '@/components/TopCompaniesSection';
 
 const CompaniesTab = () => {
   return (
